requirements/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.6
Django==4.0.4
django-crispy-forms==1.14.0
django-crum==0.7.9
django-uuidfield==0.5.0
django-widget-tweaks==1.4.12
mysqlclient==2.1.0
Pillow==9.2.0
pkg_resources==0.0.0
psycopg2==2.9.3
sqlparse==0.4.2
